edit Netplugin license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit updates the LICENSE file so that it contains the full text
of the Apache 2.0 license. This allows Licensee to successfully
identify the license type of Netplugin's codebase as Apache 2.0.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>